pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "potilastieto-backend"
version = "0.1.0"
description = "Patient information system backend"
authors = [{name = "OpenAI Codex"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "sqlmodel>=0.0.16",
    "pydantic-settings>=2.2",
    "passlib[bcrypt]>=1.7",
    "pyjwt[crypto]>=2.8",
    "python-multipart>=0.0.6",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]
exclude = ["backend.tests*"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.27",
    "ruff>=0.1.9",
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "UP", "B", "S"]
ignore = ["S101"]
